test: move fetch test to unit suite; enrich JSDoc and fix example
Follow-ups on top of the custom-fetch fix:
- Move the wrapper test from tests/management/ into the unit suite
  (src/management/tests/unit/) alongside the other wrapper tests, and
  assert on the real rules.list() return value.
- Add an explicit enriched JSDoc on the public `fetch` field.
- Fix the adjacent class-level example to use `fetch:` instead of the
  `fetcher:` callback the constructor deletes.
